Rename handles to Pre/Post Execution and ExecutionLoop
