+++ crates/revm/src/handler/handle_types.rs
@@ -1,7 +1,8 @@
 // Modules
 
-pub mod frame;
-pub mod main;
+pub mod execution_loop;
+pub mod post_execution;
+pub mod pre_execution;
 pub mod validation;
 
 // Exports
@@ -10,12 +11,15 @@ pub use validation::{
     ValidateEnvHandle, ValidateInitialTxGasHandle, ValidateTxEnvAgainstState, ValidationHandler,
 };
 
-pub use main::{
-    DeductCallerHandle, EndHandle, MainHandler, MainLoadHandle, MainLoadPrecompiles,
-    MainReturnHandle, ReimburseCallerHandle, RewardBeneficiaryHandle,
+pub use execution_loop::{
+    CreateFirstFrameHandle, ExecutionLoopHandler, FrameReturnHandle, FrameSubCallHandle,
+    FrameSubCreateHandle,
 };
 
-pub use frame::{
-    CreateFirstFrameHandle, FrameHandler, FrameReturnHandle, FrameSubCallHandle,
-    FrameSubCreateHandle,
+pub use pre_execution::{
+    DeductCallerHandle, LoadAccountsHandle, LoadPrecompilesHandle, PreExecutionHandler,
+};
+
+pub use post_execution::{
+    EndHandle, OutputHandle, PostExecutionHandler, ReimburseCallerHandle, RewardBeneficiaryHandle,
 };
