+++ crates/revm/src/handler/mainnet.rs
@@ -1,15 +1,16 @@
 //! Mainnet related handlers.
 
 mod frame;
-mod main;
+mod post_execution;
+mod pre_execution;
 mod validation;
 
 pub use frame::{
     create_first_frame, frame_return_with_refund_flag, handle_frame_return, handle_frame_sub_call,
     handle_frame_sub_create, main_frame_return,
 };
-pub use main::{
-    deduct_caller_inner, main_deduct_caller, main_end, main_load, main_load_precompiles,
-    main_reimburse_caller, main_return, main_reward_beneficiary,
+pub use post_execution::{end, output, reimburse_caller, reward_beneficiary};
+pub use pre_execution::{
+    deduct_caller_inner, main_deduct_caller, main_load, main_load_precompiles,
 };
 pub use validation::{validate_env, validate_initial_tx_gas, validate_tx_against_state};
