+++ crates/revm/src/handler/mainnet/pre_execution.rs
@@ -1,0 +1,93 @@
+//! Handles related to the main function of the EVM.
+//!
+//! They handle initial setup of the EVM, call loop and the final return of the EVM
+
+use crate::{
+    precompile::{PrecompileSpecId, Precompiles},
+    primitives::{
+        db::Database,
+        Account, EVMError, Env, Spec,
+        SpecId::{CANCUN, SHANGHAI},
+        TransactTo, U256,
+    },
+    Context,
+};
+
+/// Main precompile load
+pub fn main_load_precompiles<SPEC: Spec>() -> Precompiles {
+    Precompiles::new(PrecompileSpecId::from_spec_id(SPEC::SPEC_ID)).clone()
+}
+
+/// Main load handle
+#[inline]
+pub fn main_load<SPEC: Spec, EXT, DB: Database>(
+    context: &mut Context<EXT, DB>,
+) -> Result<(), EVMError<DB::Error>> {
+    // set journaling state flag.
+    context.evm.journaled_state.set_spec_id(SPEC::SPEC_ID);
+
+    // the L1-cost fee is only computed for Optimism non-deposit transactions.
+    #[cfg(feature = "optimism")]
+    if context.evm.env.cfg.optimism && context.evm.env.tx.optimism.source_hash.is_none() {
+        let l1_block_info = crate::optimism::L1BlockInfo::try_fetch(&mut context.evm.db)
+            .map_err(EVMError::Database)?;
+
+        // storage l1 block info for later use.
+        context.evm.l1_block_info = Some(l1_block_info);
+    }
+
+    // load coinbase
+    // EIP-3651: Warm COINBASE. Starts the `COINBASE` address warm
+    if SPEC::enabled(SHANGHAI) {
+        context
+            .evm
+            .journaled_state
+            .initial_account_load(context.evm.env.block.coinbase, &[], &mut context.evm.db)
+            .map_err(EVMError::Database)?;
+    }
+
+    context.evm.load_access_list()?;
+    Ok(())
+}
+
+/// Helper function that deducts the caller balance.
+pub fn deduct_caller_inner<SPEC: Spec>(caller_account: &mut Account, env: &Env) {
+    // Subtract gas costs from the caller's account.
+    // We need to saturate the gas cost to prevent underflow in case that `disable_balance_check` is enabled.
+    let mut gas_cost = U256::from(env.tx.gas_limit).saturating_mul(env.effective_gas_price());
+
+    // EIP-4844
+    if SPEC::enabled(CANCUN) {
+        let data_fee = env.calc_data_fee().expect("already checked");
+        gas_cost = gas_cost.saturating_add(data_fee);
+    }
+
+    // set new caller account balance.
+    caller_account.info.balance = caller_account.info.balance.saturating_sub(gas_cost);
+
+    // bump the nonce for calls. Nonce for CREATE will be bumped in `handle_create`.
+    if matches!(env.tx.transact_to, TransactTo::Call(_)) {
+        // Nonce is already checked
+        caller_account.info.nonce = caller_account.info.nonce.saturating_add(1);
+    }
+
+    // touch account so we know it is changed.
+    caller_account.mark_touch();
+}
+
+#[inline]
+pub fn main_deduct_caller<SPEC: Spec, EXT, DB: Database>(
+    context: &mut Context<EXT, DB>,
+) -> Result<(), EVMError<DB::Error>> {
+    // load caller's account.
+    let (caller_account, _) = context
+        .evm
+        .journaled_state
+        .load_account(context.evm.env.tx.caller, &mut context.evm.db)
+        .map_err(EVMError::Database)?;
+
+    // deduct gas cost from caller's account.
+    deduct_caller_inner::<SPEC>(caller_account, &context.evm.env);
+
+    Ok(())
+}
