+++ documentation/src/crates/revm/evm.md
@@ -16,11 +16,11 @@ Data structures of block and transaction can be found inside `Environment`. And 
 
 Runtime consist of list of functions from `Handler` that are called in predefined order. They are grouped by functionality on `Verification`, `Main`, `Frame` and `Instruction` functions. Verification function are related to the preverification of set `Environment` data. Main function is the main logic of the transaction execution. Frame function handles call and creates and sub calls. And Instruction functions are instruction table that executes opcodes.
 
-`Evm` execution runs **two** loops. First loop is call loop that everything starts with, it creates call frames, handles subcalls and its return outputs and call Interpreter loop to execute bytecode instructions it is handled by `FrameHandler`. Second loop is `Interpreter` loop that loops over bytecode opcodes and executes instruction from `InstructionTable`.
+`Evm` execution runs **two** loops. First loop is call loop that everything starts with, it creates call frames, handles subcalls and its return outputs and call Interpreter loop to execute bytecode instructions it is handled by `ExecutionLoopHandler`. Second loop is `Interpreter` loop that loops over bytecode opcodes and executes instruction from `InstructionTable`.
 
 First loop, the call loop, implements stack of `Frames` and it is responsible for handling sub calls and its return outputs. At the start Evm creates first `Frame` that contains `Interpreter` and starts the loop. `Interpreter` returns the `InterpreterAction` and action can be a `Return` of a call this means this interpreter finished its run or `SubCall`/`SubCreate` that means that new `Frame` needs to be created and pushed to the stack. When `Interpreter` returns `Return` action `Frame` is popped from the stack and its return value is pushed to the parent `Frame` stack. When `Interpreter` returns `SubCall`/`SubCreate` action new `Frame` is created and pushed to the stack and the loop continues. When the stack is empty the loop finishes.
 
-Second loop is `Interpreter` loop that loops over bytecode opcodes and executes instruction. It is called from the call loop and it is responsible for executing bytecode instructions. It is implemented in [`Interpreter`](../revm_interpreter/interpreter.md) crate.
+Second loop is `Interpreter` loop that loops over bytecode opcodes and executes instruction. It is called from the call loop and it is responsible for executing bytecode instructions. It is implemented in [`Interpreter`](../interpreter/interpreter.md) crate.
 
 To dive deeper into the `Evm` logic  check [`Handler`](./handler.md) documentation.
